refactored App from main
